AP_Camera: support 32 bit system IDs
Widen camera sysid storage and pass 32 bit targets via the extended
header aware send path.

+++ libraries/AP_Camera/AP_Camera_MAVLinkCamV2.h
@@ -70,7 +70,7 @@ private:
     mavlink_camera_information_t _cam_info {}; // latest camera information received from camera
     uint32_t _last_caminfo_req_ms;  // system time that CAMERA_INFORMATION was last requested (used to throttle requests)
     class GCS_MAVLINK *_link;   // link we have found the camera on. nullptr if not seen yet
-    uint8_t _sysid;             // sysid of camera
+    uint32_t _sysid;            // sysid of camera
     uint8_t _compid;            // component id of gimbal
 };
